requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
elastic-transport==8.4.0
elasticsearch==8.7.0
exceptiongroup==1.1.2
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-vision==3.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
openai==0.27.4
opencv-python==4.7.0.72
outcome==1.2.0
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PySocks==1.7.1
pytz==2023.3
requests==2.28.2
rsa==4.9
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.4
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==1.26.15
whitenoise==6.4.0
wsproto==1.2.0
yarl==1.9.1
